fix: move main.go outside of app directory
