+++ api/db.go
@@ -1,0 +1,5 @@
+package api
+
+import "database/sql"
+
+var Db *sql.DB
