+++ api/routes.go
@@ -1,18 +1,18 @@
-package main
+package api
 
 import (
-    "github.com/gorilla/mux"
+	"github.com/gorilla/mux"
 )
 
-func setupRoutes() *mux.Router {
-    router := mux.NewRouter()
+func SetupRoutes() *mux.Router {
+	router := mux.NewRouter()
 
-    // Public routes
-    router.HandleFunc("/api/users", RegisterHandler).Methods("POST")
-    router.HandleFunc("/api/auth/login", LoginHandler).Methods("POST")
+	// Public routes
+	router.HandleFunc("/api/users", RegisterHandler).Methods("POST")
+	router.HandleFunc("/api/auth/login", LoginHandler).Methods("POST")
 
-    // Protected routes
-    router.HandleFunc("/api/users/me", TokenVerifyMiddleware(UserInfoHandler)).Methods("GET")
+	// Protected routes
+	router.HandleFunc("/api/users/me", TokenVerifyMiddleware(UserInfoHandler)).Methods("GET")
 
-    return router
+	return router
 }
