+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/yourusername/go-user-service
 
-go 1.16
+go 1.23
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
